user completed and board, file building,

+++ shooney/src/main/java/com/shun/blog/dao/user/UserRepository.java
@@ -7,7 +7,7 @@ import com.shun.blog.model.user.User;
 
 public interface UserRepository {
 
-	User findById(int id);
+	User findById(Long id);
 	
 	User findByEmail(String email);
